requirements.txt
Flask
openai
langchain
python-dotenv
faiss-cpu
langchain-community